// web/src/api/generated/model/catalogueItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogueItem {
  id: string;
  title: string;
  authors: string[];
  edition: string | null;
  year: number | null;
  isbn: string | null;
  call_number: string | null;
  subjects: string[];
  copies: number;
  available: number;
  on_loan: number;
  reservations_waiting: number;
  barcodes: string[];
}